+++ lib/src/parser.rs
@@ -10,40 +10,69 @@
 
 //! The parser & parser settings
 //!
-//! The `Parser` wraps a description of a collection of available program [*options*][options] and
-//! [*command arguments*][commands], along with parser settings, and provides methods that take a
-//! list of input arguments to parse, and returns an analysis.
+//! A [`Parser`] wraps a description of a collection of available program [*options*][options] and
+//! [*command arguments*][commands], along with parser settings, and provides the parser methods
+//! that parse a given set of input arguments.
 //!
-//! The `Parser` has methods for parsing input arguments in both of the following forms:
+//! # Parsing style
+//!
+//! Parsing an argument list can be done in two different ways, depending upon your preferences or
+//! application design requirements:
+//!
+//! - “All in one” style: The [`parse`] and [`parse_os`] methods return an [`Analysis`] object,
+//!   which contains a `Vec` holding descriptions from parsing the entire argument list in one go.
+//! - “Iterative” style: The [`parse_iter`] and [`parse_iter_os`] methods return an iterator, which
+//!   returns analysis items one at a time.
+//!
+//! # Input arguments
+//!
+//! A [`Parser`] can parse input arguments in both of the following forms:
 //!
 //!  - `AsRef<str>` (`String` and `&str`)
 //!  - `AsRef<OsStr>` (`OsString` and `&OsStr`)
 //!
-//! Most programs will only want the former, however there are cases where the latter is
-//! wanted/needed.
+//! Most programs will only want the former, however there may be cases where the latter is
+//! wanted or needed. The [unicode documentation][unicode] may also be of interest here.
 //!
-//! The `Parser` also has settings to control certain aspects of parsing, for instance to choose
-//! which mode to parse options in (*standard* or *alternate*; the difference of which is discussed
-//! in the [option support documentation][options]), and whether or not to allow abbreviated *long
-//! option* name matching.
+//! Note that the [`parse`] and [`parse_iter`] methods are both `AsRef<str>` based, while
+//! [`parse_os`] and [`parse_iter_os`] are the `AsRef<OsStr>` based alternatives.
 //!
+//! # Settings
+//!
+//! A [`Parser`] has settings to control certain aspects of parsing, for instance to choose which
+//! mode to parse options in (*standard* or *alternate*; the difference of which is discussed in the
+//! [option support documentation][options]), and whether or not to allow abbreviated *long option*
+//! name matching.
+//!
+//! [`Analysis`]: ../analysis/struct.Analysis.html
+//! [`Parser`]: struct.Parser.html
+//! [`parse`]: struct.Parser.html#method.parse
+//! [`parse_iter`]: struct.Parser.html#method.parse_iter
+//! [`parse_os`]: struct.Parser.html#method.parse_os
+//! [`parse_iter_os`]: struct.Parser.html#method.parse_iter_os
 //! [commands]: ../docs/commands/index.html
 //! [options]: ../docs/options/index.html
+//! [unicode]: ../docs/unicode/index.html
 
 use std::convert::AsRef;
 use std::ffi::OsStr;
+use super::analysis::{Analysis, ItemClass};
 use super::commands::{CommandSet, CommandFlaw};
 use super::options::{OptionSet, OptionFlaw};
+
+// NB: We export this in the public API here (the `engine` and `engine_os` mods are private)
+pub use super::engine::ParseIter;
+pub use super::engine_os::ParseIterOs;
 
 /// Default abbreviation support state
 pub(crate) const ABBR_SUP_DEFAULT: bool = true;
 /// Default mode
 pub(crate) const MODE_DEFAULT: OptionsMode = OptionsMode::Standard;
 
-/// Parser
+/// The parser
 ///
-/// Holds the option set and command set used for parsing, along with parser settings, and provides
-/// the parse method for parsing an argument set.
+/// Holds the option set and command set descriptions used for parsing input arguments, along with
+/// parser settings, and provides methods for parsing.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct Parser<'r, 's: 'r> {
     /* NOTE: these have been left public to allow efficient static creation */
@@ -167,31 +196,87 @@ impl<'r, 's: 'r> Parser<'r, 's> {
         }
     }
 
-    /// Parses provided program arguments
+    /// Gives an iterator for parsing the provided program arguments
     ///
-    /// Returns an analysis describing the parsed argument list. This may include `&str` references
-    /// to strings provided in the `args` parameter and in `self`. Take note of this with respect to
+    /// Returns an iterator. Each iteration consumes one (or sometimes two) input arguments (except
+    /// with a *short option set* where one short option in the set is consumed), returning a single
+    /// analysis item.
+    ///
+    /// The returned analysis item may include `&str` references to strings provided in the `args`
+    /// parameter and/or in `self`. Take note of this with respect to object lifetimes.
+    ///
+    /// Expects `self` to be valid (see [`is_valid`](#method.is_valid)).
+    #[inline(always)]
+    pub fn parse_iter<A>(&'r self, args: &'s [A]) -> ParseIter<'r, 's, A>
+        where A: 's + AsRef<str>
+    {
+        ParseIter::new(args, self)
+    }
+
+    /// Gives an iterator for parsing the provided program arguments, given as `OsStr`
+    ///
+    /// Returns an iterator. Each iteration consumes one (or sometimes two) input arguments (except
+    /// with a *short option set* where one short option in the set is consumed), returning a single
+    /// analysis item.
+    ///
+    /// The returned analysis item may include `&str` references to strings provided in `self`,
+    /// and/or `&OsStr` to those provided in the `args` parameter. Take note of this with respect to
     /// object lifetimes.
     ///
     /// Expects `self` to be valid (see [`is_valid`](#method.is_valid)).
     #[inline(always)]
-    pub fn parse<A>(&self, args: &'s [A]) -> super::analysis::Analysis<'s, str>
-        where A: 's + AsRef<str>
-    {
-        super::engine::parse(args, self)
-    }
-
-    /// Parses provided program arguments, given as `OsStr`
-    ///
-    /// Returns an analysis describing the parsed argument list. This may include `&str` references
-    /// to strings provided in `self`, and `&OsStr` to those provided in the `args` parameter. Take
-    /// note of this with respect to object lifetimes.
-    ///
-    /// Expects `self` to be valid (see [`is_valid`](#method.is_valid)).
-    #[inline(always)]
-    pub fn parse_os<A>(&self, args: &'s [A]) -> super::analysis::Analysis<'s, OsStr>
+    pub fn parse_iter_os<A>(&'r self, args: &'s [A]) -> ParseIterOs<'r, 's, A>
         where A: 's + AsRef<OsStr>
     {
-        super::engine_os::parse(args, self)
+        ParseIterOs::new(args, self)
+    }
+
+    /// Parses the provided program arguments
+    ///
+    /// Returns an analysis describing the parsed argument list.
+    ///
+    /// The returned analysis item may include `&str` references to strings provided in the `args`
+    /// parameter and/or in `self`. Take note of this with respect to object lifetimes.
+    ///
+    /// Expects `self` to be valid (see [`is_valid`](#method.is_valid)).
+    pub fn parse<A>(&self, args: &'s [A]) -> Analysis<'s, str>
+        where A: 's + AsRef<str>
+    {
+        let mut analysis = Analysis::new(args.len());
+        let parse_iter = ParseIter::new(args, self);
+        let items = parse_iter.inspect(|item| {
+            match item {
+                ItemClass::Err(_) => analysis.error = true,
+                ItemClass::Warn(_) => analysis.warn = true,
+                ItemClass::Ok(_) => {},
+            }
+        }).collect();
+        analysis.items = items;
+        analysis
+    }
+
+    /// Parses the provided program arguments, given as `OsStr`
+    ///
+    /// Returns an analysis describing the parsed argument list.
+    ///
+    /// The returned analysis item may include `&str` references to strings provided in `self`,
+    /// and/or `&OsStr` to those provided in the `args` parameter. Take note of this with respect to
+    /// object lifetimes.
+    ///
+    /// Expects `self` to be valid (see [`is_valid`](#method.is_valid)).
+    pub fn parse_os<A>(&self, args: &'s [A]) -> Analysis<'s, OsStr>
+        where A: 's + AsRef<OsStr>
+    {
+        let mut analysis = Analysis::new(args.len());
+        let parse_iter = ParseIterOs::new(args, self);
+        let items = parse_iter.inspect(|item| {
+            match item {
+                ItemClass::Err(_) => analysis.error = true,
+                ItemClass::Warn(_) => analysis.warn = true,
+                ItemClass::Ok(_) => {},
+            }
+        }).collect();
+        analysis.items = items;
+        analysis
     }
 }